requirements.txt
python-dotenv
crewai==0.22.5
crewai[tools]
langchain-google-genai